requirements.txt

pymongo~=4.12.1
reportlab>=4.4.0
deepl>=1.22.0
tenacity>=9.1.2
psycopg2 >=2.9.10
itsdangerous
pytest
httpx
lru_cache
fastapi>=0.68.0
uvicorn>=0.34.2
python-multipart>=0.0.20
python-dotenv>=1.1.0
passlib>=1.7.4
python-jose>=3.4.0
pytesseract>=0.3.13
pillow>=11.2.1
pymupdf>=1.25.5
celery>=5.1.2
redis>=6.0.0
pypdfium2>=4.30.1
python-magic>=0.4.27
cryptography>=44.0.2
sqlalchemy>=2.0.40
fitz~=0.0.1.dev2
boto3~=1.38.7
pydantic~=2.11.4
google~=3.0.0
cachetools~=5.5.2
numpy~=2.2.5
opencv-python~=4.11.0.86
requests~=2.32.3
libretranslatepy~=2.1.4
protobuf~=6.31.0rc2
starlette~=0.46.2
pydantic-settings~=2.9.1
stripe~=12.1.0
pdfplumber~=0.11.6
arabic-reshaper~=3.0.0
openai~=1.77.0
python-bidi~=0.6.6
matplotlib~=3.10.1